[#54] Trending section added.
Also fixed [#53] - I realized the problem. I was not putting the data on the screen as soon as it was rendered. It would load data and then put it on the screen and hence the jitters. Now, it puts the data on the screen and updates it on the background.