Reverting last 3 commits which caused build error with proguard.
This reverts following commits

 - 570d26a7ae392d23065af468d814f7f78b40d48b.
 - 369e4fd311e44005e617dad15405a9717965c102
 - f6ced726969193aa115388d1bd3c0ba4d1d83ab1

It's likely that updating the gradle version caused the issue - "Invalid reference to wildcard (1, must lie between 1 and 0)" while building release APK with proguard enabled - or it could be something else. I can't find any fixes and I'm out of the loop since I did not fix this for weeks. It's a better choice just to revert the project till last stable build.