Banners support added in the app.
This commit adds support for banners in the app. Building on the idea of last commit, I integrated the new banner system.

The banner component is based on the Material Design Guidelines. It states that there should only be 1 banner.
So instead of a recyclerview, I decided to keep only one banner so that users don't feel super rushed.

The current implementation will load a list of banners, but the UI will only show one banner. Once the user interacts
with a banner, instead of bombarding the user with another banner, the app will keep a threshold of 60 mins before showing
another banner.

The 'UpdateItem' also has the capability of filtering itself out based on the environment.
Example: Redbox app support banner. If the device does not contain that app, the banner will not be shown.

This commit also adds an example for battery optimizations. I still need to build an information page for that.

Max version: Each 'UpdateItem' has a maxVersion value. If the current app version exceeds the max version, we do not show the banner.
To keep the banner visible without a max version, you set value as -1.