Fix issue with Toast - Toast shows up with delay.
The app was using a count down to show the toast for custom duration. The count down was not correctly handled
and this resulted in delay (sometimes) and crash (in Android Pie).

I have decided to remove the countdown as of now and the toast will be displayed for TOAST.LENGTH_LONG duration.
Also, removed duration seekbar in bubble settings if drawing permissions are not enabled.