Fix [#78] Support page layout and card height.
support page layout root changed to ScrollView with fillViewPort=true so that we can show the cards at the bottom.
MaxItemHeightViewPager added. This ViewPager calculates the max height among its children and sets that as self's height.
Purchase card is not wrap_content for height.

Removed MaxHeightViewPager as its no longer used.