Fixed issue of not retaining 'liked' state in collection search.
This commit fixes the issue where the 'liked' status was not persisted in the view even though it was correctly updated.

The issue was caused because the LikeStore takes care of updating all the different sub-states when MovieLiked action is dispatched and I had forgotten to add logic to change collection search page state.

This was fixed by adding logic of updating the collection search page state as well.

I wonder if this is the right approach for updated. On one hand, it becomes centralized and one place to update every state related to MovieLiked, but on the other hand this happens. Although it was easy to figure it out too.