Rating Bubble - updated to show more title details.
This commit adds 'more' details to the rating bubble including title and the year.
It also adds setting option to change back to the minimalistic view in rating bubble settings.
I also added a banner regarding this change.

Updated GA labels to see the usage of big vs small bubble.